test: moved iterdict function and added docstring to testcases
