+++ tests/conftest.py
@@ -8,17 +8,20 @@ from BALSAMIC.commands import cli
 
 @pytest.fixture
 def cli_runner():
+    """ click - cli testing """
     runner = CliRunner()
     return runner
 
 
 @pytest.fixture
 def invoke_cli(cli_runner):
+    """ invoking cli commands with options"""
     return partial(cli_runner.invoke, cli)
 
 
 @pytest.fixture(scope='session')
 def config_files():
+    """ dict: path of the config files """
     return {
         "install": "BALSAMIC/config/install.json",
         "reference": "BALSAMIC/config/reference.json",
